requirements.txt
asgiref==3.3.1
Django==3.1.5
django-filter==2.4.0
mysqlclient==2.0.3
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
xlwt==1.3.0
